requirements.txt
# Core dependencies
pyyaml>=6.0
psutil>=5.9.0
requests>=2.31.0
urllib3>=2.0.0
docker>=6.0.0

# AI providers (install the one you need)
ollama>=0.6.0  # For local Ollama models
# anthropic>=0.39.0  # Uncomment for Claude
# openai>=1.0.0  # Uncomment for OpenAI
